Rename topics/functions.py to topics/callables.py. Update CONTRIBUTING.md.
